Improvements & Make this repo only configure env

+++ first/post_download.sh
@@ -1,26 +1,31 @@
 clear
 
-source config.cfg
-
 echo "Welcome to UT4MasterServer setup helper!"
 echo ""
 echo ""
 
-# prepare partitions
-source setup/prepare_swap_partition.sh
-source setup/prepare_data_partition.sh
-mkdir /app
-
 # install all software
-source setup/install_docker.sh
-source setup/install_apache.sh
-source setup/install_certbot.sh
-source setup/install_crontab.sh
+setup/install_docker.sh
+setup/install_apache.sh
+setup/install_certbot.sh
+setup/install_crontab.sh
+setup/install_ut4ms.sh
 
-# reboot
-clear
-echo "System is going to reboot!"
-echo "After reboot, please run $(pwd)/first/post_reboot.sh script."
-sleep 1
-read -p "Press enter to continue ..."
-reboot
+# prepare partitions
+setup/prepare_swap_partition.sh
+was_swap_partition_created=$?
+setup/prepare_data_partition.sh
+was_data_partition_created=$?
+
+if [ $was_swap_partition_created -eq "1" ] -o [ $was_data_partition_created -eq "1" ]; then
+	# reboot
+	clear
+
+	findmnt --verify --verbose
+
+	echo "In order to make sure all new partitions get mounted as expected the system is going to reboot!"
+	echo "After reboot, please run \`ut4ms update\` command."
+	sleep 1
+	read -p "Press enter to continue ..."
+	reboot
+fi
